[#108] Refactor: modify index file to export each User api func
