+++ src/apis/user/index.ts
@@ -1,1 +1,4 @@
 export * from './validateNickname';
+export * from './sendAuthenticationCode';
+export * from './deleteAccountWithAuthenticationCode';
+export * from './uploadProfileImage';
